requirements.txt
langchain
langchain-community
langchain-core
langchain-text-splitters
langchain-huggingface
chromadb
pypdf
sentence-transformers
openai
tiktoken
PyQt6
PyQt6-Qt6
PyQt6-sip
numpy
pydub
mutagen
python-dotenv
imageio-ffmpeg
requests
